Frame-by-frame comparison / Make all container pydantic models
